chore: fix ruff errors
