+++ invenio_alma/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2021-2023 Graz University of Technology.
+# Copyright (C) 2021-2025 Graz University of Technology.
 #
 # invenio-alma is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -12,4 +12,9 @@ from .services import AlmaRESTService, AlmaSRUService
 
 __version__ = "0.12.0"
 
-__all__ = ("__version__", "InvenioAlma", "AlmaSRUService", "AlmaRESTService")
+__all__ = (
+    "AlmaRESTService",
+    "AlmaSRUService",
+    "InvenioAlma",
+    "__version__",
+)
