+++ invenio_alma/resources/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2023 Graz University of Technology.
+# Copyright (C) 2023-2025 Graz University of Technology.
 #
 # invenio-alma is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -11,6 +11,6 @@ from .config import AlmaResourceConfig
 from .resources import AlmaResource
 
 __all__ = (
-    "AlmaResourceConfig",
     "AlmaResource",
+    "AlmaResourceConfig",
 )
