+++ invenio_alma/services/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2022-2024 Graz University of Technology.
+# Copyright (C) 2022-2025 Graz University of Technology.
 #
 # invenio-alma is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -12,8 +12,8 @@ from .rest import AlmaRESTService
 from .sru import AlmaSRUService
 
 __all__ = (
+    "AlmaAPIError",
+    "AlmaRESTError",
     "AlmaRESTService",
     "AlmaSRUService",
-    "AlmaRESTError",
-    "AlmaAPIError",
 )
