--- invenio_alma/types.py
@@ -1,22 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright (C) 2023-2024 Graz University of Technology.
-#
-# invenio-alma is free software; you can redistribute it and/or modify it
-# under the terms of the MIT License; see LICENSE file for more details.
-
-"""Types."""
-
-from dataclasses import dataclass
-
-
-@dataclass(frozen=True)
-class Color:
-    """The class is for the output color management."""
-
-    neutral = "white"
-    error = "red"
-    warning = "yellow"
-    abort = "magenta"
-    success = "green"
-    alternate = ("blue", "cyan")
