+++ pyproject.toml
@@ -11,8 +11,8 @@ exclude = ["docs"]
 select = ["ALL"]
 
 ignore = [
-  "ANN101", "ANN102",
   "D203", "D211", "D212", "D213",
+  "E501",
   "FA102",
   "INP001",
   "G",
